package refactor, main class rename
